Add mixin examples and Quilt module
