+++ .gitignore
@@ -21,3 +21,4 @@ build
 # other
 eclipse
 run
+.DS_Store
