+++ Common/build.gradle
@@ -7,7 +7,7 @@ archivesBaseName = "${mod_id}-common-${minecraft_version}"
 dependencies {
     minecraft "com.mojang:minecraft:${minecraft_version}"
     mappings loom.officialMojangMappings()
-    implementation group: 'com.google.code.findbugs', name: 'jsr305', version: '3.0.1'
+    compileOnly group:'org.spongepowered', name:'mixin', version:'0.8.5'
 }
 
 loom {
