+++ Common/src/main/java/net/dodogang/examplemod/platform/IPlatformHelper.java
@@ -1,0 +1,27 @@
+package net.dodogang.examplemod.platform;
+
+public interface IPlatformHelper {
+    IPlatformHelper IMPL = ImplLoader.load(IPlatformHelper.class);
+
+    /**
+     * Gets the name of the current platform
+     *
+     * @return The name of the current platform.
+     */
+    String getPlatformName();
+
+    /**
+     * Checks if a mod with the given id is loaded.
+     *
+     * @param modId The mod to check if it is loaded.
+     * @return True if the mod is loaded, false otherwise.
+     */
+    boolean isModLoaded(String modId);
+
+    /**
+     * Check if the game is currently in a development environment.
+     *
+     * @return True if in a development environment, false otherwise.
+     */
+    boolean isDevelopmentEnvironment();
+}
