+++ Common/src/main/java/net/dodogang/examplemod/platform/ImplLoader.java
@@ -1,0 +1,13 @@
+package net.dodogang.examplemod.platform;
+
+
+
+import java.util.ServiceLoader;
+
+public class ImplLoader {
+    public static <T> T load(Class<T> clazz) {
+        return ServiceLoader.load(clazz)
+                .findFirst()
+                .orElseThrow(() -> new NullPointerException("Failed to load service for " + clazz.getName()));
+    }
+}
