--- Fabric/src/main/java/com/example/examplemod/platform/FabricPlatformHelper.java
@@ -1,24 +1,0 @@
-package com.example.examplemod.platform;
-
-import com.example.examplemod.platform.services.IPlatformHelper;
-import net.fabricmc.loader.api.FabricLoader;
-
-public class FabricPlatformHelper implements IPlatformHelper {
-
-    @Override
-    public String getPlatformName() {
-        return "Fabric";
-    }
-
-    @Override
-    public boolean isModLoaded(String modId) {
-
-        return FabricLoader.getInstance().isModLoaded(modId);
-    }
-
-    @Override
-    public boolean isDevelopmentEnvironment() {
-
-        return FabricLoader.getInstance().isDevelopmentEnvironment();
-    }
-}
