+++ Fabric/src/main/java/net/dodogang/examplemod/platform/FabricPlatformHelper.java
@@ -1,0 +1,20 @@
+package net.dodogang.examplemod.platform;
+
+import net.fabricmc.loader.api.FabricLoader;
+
+public class FabricPlatformHelper implements IPlatformHelper {
+    @Override
+    public String getPlatformName() {
+        return "Fabric";
+    }
+
+    @Override
+    public boolean isModLoaded(String modId) {
+        return FabricLoader.getInstance().isModLoaded(modId);
+    }
+
+    @Override
+    public boolean isDevelopmentEnvironment() {
+        return FabricLoader.getInstance().isDevelopmentEnvironment();
+    }
+}
