+++ Fabric/src/main/resources/META-INF/services/net.dodogang.examplemod.platform.IPlatformHelper
@@ -1,0 +1,1 @@
+net.dodogang.examplemod.platform.FabricPlatformHelper
