--- Fabric/src/main/resources/modid.mixins.json
@@ -1,15 +1,0 @@
-{
-    "required": true,
-    "minVersion": "0.8",
-    "package": "com.example.examplemod.mixin",
-    "compatibilityLevel": "JAVA_17",
-    "mixins": [
-    ],
-    "client": [
-      "ExampleMixin"
-    ],
-    "injectors": {
-      "defaultRequire": 1
-    }
-  }
-
