+++ Forge/build.gradle
@@ -1,18 +1,29 @@
 buildscript {
     repositories {
         maven { url = 'https://maven.minecraftforge.net' }
+        maven { url = 'https://repo.spongepowered.org/repository/maven-public/' }
         mavenCentral()
     }
     dependencies {
         classpath group: 'net.minecraftforge.gradle', name: 'ForgeGradle', version: '5.1.+', changing: true
+        classpath 'org.spongepowered:mixingradle:0.7-SNAPSHOT'
     }
 }
 apply plugin: 'java'
 apply plugin: 'net.minecraftforge.gradle'
 apply plugin: 'eclipse'
-apply plugin: 'maven-publish'
+apply plugin: 'org.spongepowered.mixin'
 
 archivesBaseName = "${mod_id}-forge-${minecraft_version}"
+
+mixin {
+    add project(":Common").sourceSets.main, "${mod_id}.refmap.json"
+
+    add sourceSets.main, "${mod_id}-forge.refmap.json"
+
+    config "${mod_id}.mixins.json"
+    config "${mod_id}-forge.mixins.json"
+}
 
 minecraft {
     mappings channel: 'official', version: minecraft_version
@@ -31,6 +42,7 @@ minecraft {
             taskName 'Client'
             property 'mixin.env.remapRefMap', 'true'
             property 'mixin.env.refMapRemappingFile', "${projectDir}/build/createSrgToMcp/output.srg"
+            args "-mixin.config=${mod_id}.mixins.json", "-mixin.config=${mod_id}-forge.mixins.json"
             mods {
                 modClientRun {
                     source sourceSets.main
@@ -45,6 +57,7 @@ minecraft {
             taskName 'Server'
             property 'mixin.env.remapRefMap', 'true'
             property 'mixin.env.refMapRemappingFile', "${projectDir}/build/createSrgToMcp/output.srg"
+            args "-mixin.config=${mod_id}.mixins.json", "-mixin.config=${mod_id}-forge.mixins.json"
             mods {
                 modServerRun {
                     source sourceSets.main
@@ -60,6 +73,7 @@ minecraft {
             taskName 'Data'
             property 'mixin.env.remapRefMap', 'true'
             property 'mixin.env.refMapRemappingFile', "${projectDir}/build/createSrgToMcp/output.srg"
+            args "-mixin.config=${mod_id}.mixins.json", "-mixin.config=${mod_id}-forge.mixins.json"
             mods {
                 modDataRun {
                     source sourceSets.main
@@ -74,6 +88,7 @@ sourceSets.main.resources.srcDir 'src/generated/resources'
 
 dependencies {
     minecraft "net.minecraftforge:forge:${minecraft_version}-${forge_version}"
+    annotationProcessor 'org.spongepowered:mixin:0.8.4-SNAPSHOT:processor'
     compileOnly project(path: ':Common', configuration: 'namedElements')
 }
 
@@ -83,6 +98,14 @@ tasks.withType(JavaCompile) {
 
 processResources {
     from project(':Common').sourceSets.main.resources
+
+    filesMatching("${mod_id}.mixins.json") {
+        expand "refmap_target": "${mod_id}."
+    }
+
+    filesMatching("${mod_id}-forge.mixins.json") {
+        expand "refmap_target": "${mod_id}-"
+    }
 }
 
 jar.finalizedBy('reobfJar')
