--- Forge/src/main/java/com/example/examplemod/platform/ForgePlatformHelper.java
@@ -1,26 +1,0 @@
-package com.example.examplemod.platform;
-
-import com.example.examplemod.platform.services.IPlatformHelper;
-import net.minecraftforge.fml.ModList;
-import net.minecraftforge.fml.loading.FMLLoader;
-
-public class ForgePlatformHelper implements IPlatformHelper {
-
-    @Override
-    public String getPlatformName() {
-
-        return "Forge";
-    }
-
-    @Override
-    public boolean isModLoaded(String modId) {
-
-        return ModList.get().isLoaded(modId);
-    }
-
-    @Override
-    public boolean isDevelopmentEnvironment() {
-
-        return !FMLLoader.isProduction();
-    }
-}
