+++ Forge/src/main/java/net/dodogang/examplemod/platform/ForgePlatformHelper.java
@@ -1,0 +1,21 @@
+package net.dodogang.examplemod.platform;
+
+import net.minecraftforge.fml.ModList;
+import net.minecraftforge.fml.loading.FMLLoader;
+
+public class ForgePlatformHelper implements IPlatformHelper {
+    @Override
+    public String getPlatformName() {
+        return "Forge";
+    }
+
+    @Override
+    public boolean isModLoaded(String modId) {
+        return ModList.get().isLoaded(modId);
+    }
+
+    @Override
+    public boolean isDevelopmentEnvironment() {
+        return !FMLLoader.isProduction();
+    }
+}
