+++ Forge/src/main/resources/examplemod-forge.mixins.json
@@ -1,0 +1,18 @@
+{
+    "required": true,
+    "minVersion": "0.8",
+    "package": "net.dodogang.examplemod.mixin",
+    "compatibilityLevel": "JAVA_17",
+    "mixins": [
+    ],
+    "client": [
+      "client.ExampleForgeMixin"
+    ],
+    "server": [
+    ],
+    "injectors": {
+      "defaultRequire": 1
+    },
+    "refmap": "${refmap_target}forge.refmap.json"
+  }
+
