+++ Quilt/src/main/java/net/dodogang/examplemod/mixin/client/ExampleQuiltMixin.java
@@ -1,0 +1,16 @@
+package net.dodogang.examplemod.mixin.client;
+
+import net.dodogang.examplemod.ExampleMod;
+import net.minecraft.client.gui.screens.TitleScreen;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(TitleScreen.class)
+public class ExampleQuiltMixin {
+    @Inject(at = @At("HEAD"), method = "init()V")
+    private void init(CallbackInfo info) {
+        ExampleMod.LOGGER.info("This line is printed by a mixin from Quilt!");
+    }
+}
