+++ Quilt/src/main/java/net/dodogang/examplemod/platform/QuiltPlatformHelper.java
@@ -1,0 +1,20 @@
+package net.dodogang.examplemod.platform;
+
+import org.quiltmc.loader.api.QuiltLoader;
+
+public class QuiltPlatformHelper implements IPlatformHelper {
+    @Override
+    public String getPlatformName() {
+        return "Quilt";
+    }
+
+    @Override
+    public boolean isModLoaded(String modId) {
+        return QuiltLoader.isModLoaded(modId);
+    }
+
+    @Override
+    public boolean isDevelopmentEnvironment() {
+        return QuiltLoader.isDevelopmentEnvironment();
+    }
+}
