+++ settings.gradle
@@ -24,4 +24,4 @@ pluginManagement {
 }
 
 rootProject.name = 'MultiLoader'
-include('Common', 'Fabric', 'Forge')
+include('Common', 'Fabric', 'Forge', 'Quilt')
